added scripts to build docker images and publish it using Travis

+++ kafka/docker/scripts/common.sh
@@ -14,11 +14,13 @@
 
 show_help() {
 cat << EOF
-Usage: ${0##*/} [-h] [-p] [-r]
+Usage: ${0##*/} [-h] [-p] [-b] [-r] [--release]
 
     -h, --help      display this help and exit.
     -p, --package   maven package.
+    -b, --build     docker build.
     -r, --run       run docker image.
+    --release       release docker image
 EOF
 }
 
@@ -33,8 +35,14 @@ while :; do
         -p|--package)
             package=1
             ;;
+        -b|--build)
+            build=1
+            ;;
         -r|--run)
             run=1
+            ;;
+        --release)
+            release=1
             ;;
         --)              # End of all options.
             shift
@@ -57,32 +65,41 @@ if [ "$package" == "1" ]; then
     cd kafka/docker/scripts
 fi
 
-cd ..
-pwd
 
-echo 'Building...'
-rm -rf tmp # Let's keep the temp files for easy validating until next run
-mkdir tmp
-cp template/* tmp
+if [ "$build" == "1" ]; then
+    cd ..
+    pwd
 
-# simple templating based off {{{var}}} not for advanced usage
-echo -e "s/{{{APP_NAME}}}/$APP_NAME/\n\
-s/{{{DOTTED_APP_NAME}}}/$DOTTED_APP_NAME/\n\
-s/{{{MAIN_CLASS}}}/$MAIN_CLASS/\n" > tmp/vars.sed
+    echo 'Building...'
+    rm -rf tmp # Let's keep the temp files for easy validating until next run
+    mkdir tmp
+    cp template/* tmp
 
-sed -i '' -f tmp/vars.sed tmp/Dockerfile
-sed -i '' -f tmp/vars.sed tmp/jmxtrans-agent.xml
-sed -i '' -f tmp/vars.sed tmp/start-app.sh
+    # simple templating based off {{{var}}} not for advanced usage
+    echo -e "s/{{{APP_NAME}}}/$APP_NAME/\n\
+    s/{{{DOTTED_APP_NAME}}}/$DOTTED_APP_NAME/\n\
+    s/{{{MAIN_CLASS}}}/$MAIN_CLASS/\n" > tmp/vars.sed
+
+    sed -i '' -f tmp/vars.sed tmp/Dockerfile
+    sed -i '' -f tmp/vars.sed tmp/jmxtrans-agent.xml
+    sed -i '' -f tmp/vars.sed tmp/start-app.sh
 
 
-# copy .. seems to fail in docker build
-cd ..
-pwd
+    # copy .. seems to fail in docker build
+    cd ..
+    pwd
 
-docker build -f docker/tmp/Dockerfile -t "$APP_NAME" .
-
+    docker build -f docker/tmp/Dockerfile -t "$APP_NAME" .
+fi
 
 if [ "$run" == "1" ]; then
     echo 'Running...'
     docker run "$APP_NAME"
+fi
+
+if [ "$release" == "1" ]; then
+    echo "Releasing"
+    export DOCKER_ORG="expediadotcom"
+    export DOCKER_IMAGE_NAME=$APP_NAME
+    ../../../scripts/publish-to-docker-hub.sh
 fi
